Labs:   Create an Extensible Super Class   Magic Methods   Abstract Classes

+++ index.php
@@ -8,10 +8,10 @@ spl_autoload_register(
     }
 );
 
-use \Server\VPS;
+use \Machine\VPS;
 
 $vps = new VPS('newvps', '8GB', '16GB');
-var_dump($vps);
+echo "toString: [".$vps."]\n";
 
 if ( $vps->create(true) ) {
     echo "VPS created\n";
@@ -20,7 +20,23 @@ else {
     echo "VPS couldn't be created\n";
 }
 
+echo "Starting the VPS...\n";
+if ( ! $vps->start() ) {
+    echo "Error starting VPS!\n";
+    exit(1);
+}
+
 sleep(1);
+
+echo "Generating report...\n";
+$report = $vps->getReport();
+echo "REPORT: ".$report->generate()."\n";
+
+echo "Stopping the VPS...\n";
+if ( ! $vps->stop() ) {
+    echo "Error stopping VPS!\n";
+    exit(2);
+}
 
 if ( $vps->delete() ) {
     echo "VPS removed\n";
@@ -29,3 +45,6 @@ else {
     echo "VPS couldn't be removed\n";
 }
 
+$vps->opticalDrive = 'Pioneer';
+
+exit(0);
